ai-service/requirements.txt
# # --- Web Server ---
# fastapi
# uvicorn
# python-multipart

# # --- AI Frameworks (Bản chỉ chạy CPU để giảm dung lượng) ---
# --extra-index-url https://download.pytorch.org/whl/cpu
# torch==2.5.1
# torchvision==0.20.1

# ultralytics==8.1.0

# # Bản PaddlePaddle chỉ dành cho CPU
# paddlepaddle==2.6.2

# paddleocr==2.8.1

# # --- Compatibility & Utilities ---
# numpy==1.26.4
# opencv-python-headless==4.9.0.80

# --- Web Server ---
fastapi
uvicorn
python-multipart

# --- AI Inference (Siêu nhẹ, không dùng PyTorch/Paddle) ---
onnxruntime
rapidocr-onnxruntime

# --- Utilities ---
Pillow==10.2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
